# user-service/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "https://curso-automacao-user-service.herokuapp.com",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/v1/users/save": {
            "put": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "save",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/update": {
            "post": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "username": {
                        "type": "string"
                    },
                    "userpass": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "string"
                    }
                }
            }
        }
    }
}